Features Resort Property Admin Fxns Resort Property Types Admin Fxns SQL Script addition for neede vehicle status types Update to sp_delete_resort_property_type_by_id to use an int Refactor for role checking for admin functions on resort property type and ressort property

+++ MillennialResortManager/LogicLayer/ResortPropertyManager.cs
@@ -124,10 +124,13 @@ namespace LogicLayer
         /// Deletes Resort Property From database
         /// </summary>
         /// <param name="id"></param>
-        public void DeleteResortProperty(string id)
+        public void DeleteResortProperty(int id, Employee employee)
         {
             try
             {
+                if(!employee.HasRoles(out var errorStr, "Admin"))
+                    throw new ApplicationException(errorStr);
+
                 _resortPropertyAccessor.DeleteResortProperty(id);
             }
             catch (Exception)
